requirements_minimal.txt
numpy
scikit-learn
opencv-python
matplotlib
streamlit
joblib
Pillow
plotly
pandas